+++ src/lang/en.ts
@@ -75,6 +75,7 @@ export const translation: Translation = {
             failedTarget: 'Could not deploy {0:trim,surround} to ${1:trim} on save: {2}',
         },
         operations: {
+            noFunctionInScript: "The function {0:trim,surround} was not found in {1:trim,surround}!",
             open: 'Opening {0:trim,surround}... ',
             unknownType: 'UNKNOWN TYPE: {0:trim,surround}',
         },